[uniffi] Create dedicated test for custom storage
This lets us make the “simple scenario” integration test, well, simple
again.
