+++ mls-rs-uniffi/tests/scenarios.rs
@@ -44,6 +44,7 @@ macro_rules! generate_python_tests {
 
 generate_python_tests!(generate_signature_keypair, None);
 generate_python_tests!(client_config_default_sync, None);
+generate_python_tests!(custom_storage_sync, None);
 
 // TODO(mulmarta): it'll break if we use async trait which will be
 // supported in the next UniFFI release
